+++ Sources/VComponents/Components/Item Pickers/Wheel Picker/VWheelPicker.swift
@@ -6,11 +6,12 @@
 //
 
 import SwiftUI
+import VCore
 
 // MARK: - V Wheel Picker
 /// Item picker component that selects from a set of mutually exclusive values, and displays their representative content in a scrollable wheel.
 ///
-/// Component can be initialized with data, row titles, `PickableEnumeration`, or `PickableTitledItem`.
+/// Component can be initialized with data, row titles, `HashableEnumeration`, or `StringRepresentableHashableEnumeration`.
 ///
 /// Best suited for `5`+ items.
 ///
@@ -138,49 +139,49 @@ public struct VWheelPicker<Data, Content>: View
         self.content = .titles(titles: rowTitles)
     }
     
-    // MARK: Initializers - Pickable Enumeration & Pickable Titled Enumeration
-    /// Initializes component with `PickableEnumeration` and row content.
-    public init<PickableItem>(
+    // MARK: Initializers - Hashable Enumeration & String Representable Hashable Enumeration
+    /// Initializes component with `HashableEnumeration` and row content.
+    public init<T>(
         uiModel: VWheelPickerUIModel = .init(),
-        selection: Binding<PickableItem>,
+        selection: Binding<T>,
         headerTitle: String? = nil,
         footerTitle: String? = nil,
-        @ViewBuilder content: @escaping (PickableItem) -> Content
+        @ViewBuilder content: @escaping (T) -> Content
     )
         where
-            Data == Array<PickableItem>,
-            PickableItem: PickableEnumeration
+            Data == Array<T>,
+            T: HashableEnumeration
     {
         self.uiModel = uiModel
         self._selectedIndex = .init(
-            get: { Array(PickableItem.allCases).firstIndex(of: selection.wrappedValue)! }, // fatalError
-            set: { selection.wrappedValue = Array(PickableItem.allCases)[$0] }
+            get: { Array(T.allCases).firstIndex(of: selection.wrappedValue)! }, // fatalError
+            set: { selection.wrappedValue = Array(T.allCases)[$0] }
         )
         self.headerTitle = headerTitle
         self.footerTitle = footerTitle
-        self.content = .custom(data: Array(PickableItem.allCases), content: content)
+        self.content = .custom(data: Array(T.allCases), content: content)
     }
     
-    /// Initializes component with `PickableTitledEnumeration`.
-    public init<PickableItem>(
+    /// Initializes component with `StringRepresentableHashableEnumeration`.
+    public init<T>(
         uiModel: VWheelPickerUIModel = .init(),
-        selection: Binding<PickableItem>,
+        selection: Binding<T>,
         headerTitle: String? = nil,
         footerTitle: String? = nil
     )
         where
             Data == Array<Never>,
             Content == Never,
-            PickableItem: PickableTitledEnumeration
+            T: StringRepresentableHashableEnumeration
     {
         self.uiModel = uiModel
         self._selectedIndex = .init(
-            get: { Array(PickableItem.allCases).firstIndex(of: selection.wrappedValue)! }, // fatalError
-            set: { selection.wrappedValue = Array(PickableItem.allCases)[$0] }
+            get: { Array(T.allCases).firstIndex(of: selection.wrappedValue)! }, // fatalError
+            set: { selection.wrappedValue = Array(T.allCases)[$0] }
         )
         self.headerTitle = headerTitle
         self.footerTitle = footerTitle
-        self.content = .titles(titles: Array(PickableItem.allCases).map { $0.pickerTitle })
+        self.content = .titles(titles: Array(T.allCases).map { $0.stringRepresentation })
     }
 
     // MARK: Body